requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.7
aiohttp==3.10.3
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.6.2
boto3==1.28.15
botocore==1.31.15
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.6
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
coloredlogs==15.0.1
comm @ file:///C:/b/abs_67a8058udb/croot/comm_1709322909844/work
contourpy==1.2.0
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
effdet==0.4.1
emoji==2.12.1
exceptiongroup @ file:///C:/b/abs_c5h1o1_b5b/croot/exceptiongroup_1706031441653/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.112.0
ffmpy==0.4.0
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.17.1
google-api-python-client==2.140.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-vision==3.7.4
google-generativeai==0.7.2
googleapis-common-protos==1.63.0
gradio==4.41.0
gradio_client==1.3.0
greenlet==3.0.3
groq==0.9.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.6
importlib_metadata==8.0.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel @ file:///C:/b/abs_c2u94kxcy6/croot/ipykernel_1705933907920/work
ipython @ file:///C:/b/abs_7dmlwmkafn/croot/ipython_1718287997280/work
itsdangerous==2.1.2
jedi @ file:///C:/b/abs_1b8kmj7rrm/croot/jedi_1721058359741/work
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///C:/b/abs_a6h3c8hfdq/croot/jupyter_client_1699455939372/work
jupyter_core @ file:///C:/b/abs_beftpbuevw/croot/jupyter_core_1718818307097/work
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.14
langchain-chroma==0.1.2
langchain-community==0.2.12
langchain-core==0.2.33
langchain-google-genai==1.0.9
langchain-groq==0.1.9
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.99
layoutparser==0.3.4
llvmlite==0.42.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.8.3
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934094726/work
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.4.0
nest-asyncio @ file:///C:/b/abs_65d6lblmoi/croot/nest-asyncio_1708532721305/work
networkx==3.1
ngrok==1.2.0
nltk==3.8.1
numba==0.59.0
numpy==1.25.1
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.0
openai==1.41.0
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
opencv-python==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
ordered-set==4.1.0
orjson==3.10.7
overrides==7.7.0
packaging==24.0
pandas==2.0.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.3
pikepdf==9.1.1
pillow==10.2.0
pillow_heif==0.18.0
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
portalocker==2.10.1
posthog==3.5.0
prompt-toolkit @ file:///C:/b/abs_68uwr58ed1/croot/prompt-toolkit_1704404394082/work
proto-plus==1.23.0
protobuf==4.25.3
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.1
pydub==0.25.1
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
pymongo==4.6.2
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyngrok==7.1.5
pyparsing==3.1.2
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.9
pytz==2023.3
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_89aq69t0up/croot/pyzmq_1705605705281/work
rapidfuzz==3.1.2
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruff==0.5.7
s3transfer==0.6.1
safetensors==0.4.4
scipy==1.14.0
semantic-version==2.10.0
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.37.2
streamlit==1.37.1
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.0.1
torchvision==0.19.0
tornado @ file:///C:/b/abs_7bua0304mj/croot/tornado_1718740122405/work
tqdm==4.66.2
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
transformers==4.44.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
unstructured==0.15.5
unstructured-client==0.25.5
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.23.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.3.6
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.0
